refactor(view-model): reorder directories for better organization
